requirements.txt
APScheduler
Flask
Flask-APScheduler
Flask-RESTful
Flask-SQLAlchemy
gunicorn
pycryptodome
pytest
pytest-flask
pytest-mock
requests
